Added the first files of the Mathblock Service
This service is not ready yet.
For now the algorithm to view the mathematical expressions is added, as well as some HTTP Polling logic to make this feature work in harmony with the DIASTEMA Orchestrator.
Currently I am working on making the subtraction operation to work for this feature. After this operation all the other ones will soon be ready as well.